Add scripts from current develop branch
This commit introduces several new scripts for more sophisticated peak filtering and metadata normalization. It includes functions for filtering peaks based on their intensity and mass-to-charge ratio, as well as functions for normalizing ionization mode, instruments and resolutions, adducts, retention time, and predicted values amongst others. In addition, the commit incorporates the ability to calculate molecular properties based on unique InChi and SMILES in a given DataFrame. And scripts to repair molecular descriptors where necessary.