Add old_spectrum to return statements in msp_normalizer.py
The return statements in msp_normalizer.py have been updated to include the old_spectrum. This change allows the older version of the spectrum to be returned along with the other values when certain conditions are not satisfied or in case of failure in peak list preparation and normalization step.