Add function to structure spectrum traces
A new function called 'structure_traces_one' was introduced to the TRACES.py script. This function processes a list of spectra traces and organizes them into a list of dictionaries. This newly structured list is then fed into the MAIN.py script. It ensures that the data from the spectra traces are better structured for further processing.