Refactor deletion flag name in normalizer script
The name of the flag column used for marking rows to be deleted in the mols_calculation script has been refactored from 'will_be_deleted' to 'will_be_deleted_because_of_RDkit'. This change aims to provide a clearer context for its usage and the reason behind the deletion of certain rows.