Add structure_traces_two method and update MAIN.py
A new function named structure_traces_two has been added to the TRACES.py script. This function takes a list of spectrums, merges them, concatenates with a DataFrame of deleted spectrums, and saves the output as an Excel file. Additionally, in MAIN.py, a line of code that was printing the first element of the spectrum_list_TRACES array has been commented out and a call to the newly added structure_traces_two function has been added after the mols_derivation_and_calculation method call.