Update condition and dataframe handling in TRACES.py, add print in MAIN.py
The condition for appending to TRACES_DICT_LIST in TRACES.py is updated to check for 'None' instead of 'not None'. DataFrame creation from temp_spectrums dictionary has been modified to use a list-comprehension ensuring every value is a list and to use the default index. In MAIN.py, a print statement is added to display the first item of spectrum_list_TRACES.