Refactor spectrum traces processing for concurrency
Refactored the spectrum traces processing script to make use of concurrent processing, which allows multiple spectra to be cleaned at once. This was achieved by chunking the list of spectra and using the ThreadPoolExecutor from the concurrent.futures module. Progress tracking was also added to the cleaning operation to provide visual feedback to the user.