Refactor fraghubid_generator.py and remove unnecessary patterns
The commit removes the regular expressions that were used to parse the 'inchikey' and 'peak_list' from the spectrum data. Now, the script directly accesses the 'PEAKS_LIST' from the spectrum data and no longer needs to extract the 'inchikey'. It leads to clearer, less complex code and eliminates the requirement for regex patterns and their global declarations.